refactor(react-property): update build output and filenames
The html and svg properties configs are no longer under a
specific directory but kept flat under the `lib/` directory:

```sh
$ cd packages/react-property/
$ tree lib/
lib/
├── HTMLDOMPropertyConfig.json
├── SVGDOMPropertyConfig.json
├── html
│   └── index.js
├── injection.json
├── is-custom-attribute.js
└── svg
    └── index.js

2 directories, 6 files
```
