+++ packages/react-property/src/svg/index.js
@@ -1,4 +1,4 @@
-var properties = require('./properties');
+var properties = require('../SVGDOMPropertyConfig');
 var injection = require('../injection');
 
 var MUST_USE_PROPERTY = injection.MUST_USE_PROPERTY;
